Introduce Route list activity. move 'Clear all' button. Details view is opened from route list.
